+++ plugins/dialog/plugin.js
@@ -3224,12 +3224,14 @@ CKEDITOR.DIALOG_STATE_BUSY = 2;
 			 */
 			functions: function() {
 				var args = arguments;
-				return function() {
+				return function( value ) {
 					// It's important for validate functions to be able to accept the value
 					// as argument in addition to this.getValue(), so that it is possible to
 					// combine validate functions together to make more sophisticated
 					// validators.
-					var value = this && this.getValue ? this.getValue() : args[ 0 ];
+					if ( this && this.getValue ) {
+						value = this.getValue();
+					}
 
 					var msg,
 						relation = CKEDITOR.VALIDATE_AND,
@@ -3237,10 +3239,11 @@ CKEDITOR.DIALOG_STATE_BUSY = 2;
 						i;
 
 					for ( i = 0; i < args.length; i++ ) {
-						if ( typeof args[ i ] == 'function' )
+						if ( typeof args[ i ] == 'function' ) {
 							functions.push( args[ i ] );
-						else
+						} else {
 							break;
+						}
 					}
 
 					if ( i < args.length && typeof args[ i ] == 'string' ) {
@@ -3279,8 +3282,8 @@ CKEDITOR.DIALOG_STATE_BUSY = 2;
 			 * Checks if a dialog UI element value meets the regex condition.
 			 *
 			 * ```javascript
-			 * CKEDITOR.dialog.validate.regex( 'error!', /^\d*$/ )( '123' ) // true
-			 * CKEDITOR.dialog.validate.regex( 'error!' )( '123.321' ) // error!
+			 * CKEDITOR.dialog.validate.regex( /^\d*$/, 'error!' )( '123' ) // true
+			 * CKEDITOR.dialog.validate.regex( /^\d*$/, 'error!' )( '123.321' ) // error!
 			 * ```
 			 *
 			 * @param {RegExp} regex Regular expression used to validate the value.
@@ -3288,14 +3291,9 @@ CKEDITOR.DIALOG_STATE_BUSY = 2;
 			 * @returns {Function} Validation function.
 			 */
 			regex: function( regex, msg ) {
-				/*
-				 * Can be greatly shortened by deriving from functions validator if code size
-				 * turns out to be more important than performance.
-				 */
-				return function() {
-					var value = this && this.getValue ? this.getValue() : arguments[ 0 ];
-					return !regex.test( value ) ? msg : true;
-				};
+				return this.functions( function( val ) {
+					return regex.test( val );
+				}, msg );
 			},
 
 			/**
@@ -3310,14 +3308,12 @@ CKEDITOR.DIALOG_STATE_BUSY = 2;
 			 * @returns {Function} Validation function.
 			 */
 			notEmpty: function( msg ) {
-				var trimCharacters = '\\u0020\\u00a0\\u1680\\u202f\\u205f\\u3000\\u2000-\\u200a\\s',
-					trimRegex = new RegExp( '^[' + trimCharacters + ']+|[' + trimCharacters + ']+$', 'g' );
+				return this.functions( function( val ) {
+					var trimCharacters = '\\u0020\\u00a0\\u1680\\u202f\\u205f\\u3000\\u2000-\\u200a\\s',
+						trimRegex = new RegExp( '^[' + trimCharacters + ']+|[' + trimCharacters + ']+$', 'g' );
 
-				return function() {
-					var value = this && this.getValue ? this.getValue() : arguments[ 0 ];
-
-					return value.replace( trimRegex, '' ).length > 0 || msg;
-				};
+					return val.replace( trimRegex, '' ).length > 0;
+				}, msg );
 			},
 
 			/**